pyproject.toml
[project]
name = "ai-invoice-auditor"
version = "0.1.0"
description = "AI-powered invoice auditing system with agentic RAG"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10,<3.0",
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "pytest>=8.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "rapidfuzz>=3.14.3",
    "filetype>=1.2.0",
    "python-docx>=1.2.0",
    "pytesseract>=0.3.13",
    "pdfplumber>=0.11.9",
    "langgraph>=0.2",
    "fastmcp==2.14.5",
    "jinja2>=3.1",
    "chromadb>=1.0.0",
    "langchain-chroma>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langfuse>=4.0.0",
    "streamlit>=1.55.0",
    "langchain>=1.2.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
